Splits models.py into models/*.py
