+++ src/hueb/apps/hueb20/models/__init__.py
@@ -1,0 +1,11 @@
+LATEIN = "LATEIN"
+LIDOS = "LIDOS "
+LEGACY = "LEGACY"
+HUEB20 = "HUEB20"
+
+HUEB_APPLICATIONS = [
+    (LATEIN, "HÜB Latein Datensatz"),
+    (LIDOS, "HÜB Lidos Datensatz"),
+    (LEGACY, "HÜB Basis Datensatz"),
+    (HUEB20, "HÜB 2020 Datensatz"),
+]
