+++ src/hueb/apps/hueb20/models/comment.py
@@ -1,0 +1,43 @@
+from django.db import models
+from django.template.defaultfilters import escape
+from hueb.apps.hueb20.models import HUEB20, HUEB_APPLICATIONS
+
+from .culturalCircle import CulturalCircle
+from .document import Document
+from .person import Person
+
+
+class Comment(models.Model):
+    id = models.BigAutoField(primary_key=True)
+    text = models.TextField(blank=True, null=True)
+    external = models.BooleanField(
+        default=False,
+        help_text="External comments will be displayed on the public website.",
+    )
+    person = models.ForeignKey(
+        Person,
+        on_delete=models.CASCADE,
+        null=True,
+        blank=True,
+        related_name="person_comment",
+    )
+    document = models.ForeignKey(
+        Document,
+        on_delete=models.CASCADE,
+        null=True,
+        blank=True,
+        related_name="document_comment",
+    )
+    cultural_circle = models.ForeignKey(
+        CulturalCircle,
+        on_delete=models.CASCADE,
+        null=True,
+        blank=True,
+        related_name="cultural_circle_comment",
+    )
+    app = models.CharField(max_length=6, choices=HUEB_APPLICATIONS, default=HUEB20)
+
+    def __str__(self):
+        if self.text is None:
+            return " "
+        return escape(self.text)
