+++ src/hueb/apps/hueb20/models/language.py
@@ -1,0 +1,21 @@
+import hueb.apps.hueb_legacy_latein.models as Legacy
+from django.db import models
+from hueb.apps.hueb20.models import HUEB20, HUEB_APPLICATIONS
+
+
+class Language(models.Model):
+    id = models.BigAutoField(primary_key=True)
+    language = models.CharField(max_length=255)
+    app = models.CharField(max_length=6, choices=HUEB_APPLICATIONS, default=HUEB20)
+    language_ref = models.OneToOneField(
+        Legacy.Language,
+        on_delete=models.DO_NOTHING,
+        null=True,
+        blank=True,
+        related_name="language_ref",
+    )
+
+    def __str__(self):
+        if self.language is None:
+            return " "
+        return self.language
